Improve MMC3 IRQ logic

+++ src/Hardware/h_MMC3.h
@@ -6,7 +6,7 @@
 typedef	struct	MMC3
 {
 	u8 IRQenabled, IRQcounter, IRQlatch, IRQreload;
-	int IRQaddr;
+	u8 IRQaddr;
 	u8 Cmd;
 	u8 PRG[4];
 	u8 CHR[8];
